Introduce the Record trait enabling path substitution
The `inject_values_into_template` method has been extracted into a
`trait Record` making the feature usable not only with
GenericCommandState (a workflow specific type) but with any types
implementing the `Record` trait. i.e. the ability to extract a json
value given a path.
