+++ crates/core/tedge_api/src/workflow/mod.rs
@@ -10,6 +10,7 @@ use crate::mqtt_topics::EntityTopicId;
 use crate::mqtt_topics::MqttSchema;
 use crate::mqtt_topics::OperationType;
 use crate::script::ShellScript;
+use crate::substitution::Record;
 use ::log::info;
 pub use error::*;
 pub use handlers::*;
@@ -345,19 +346,17 @@ impl OperationWorkflow {
 impl OperationAction {
     pub fn inject_state(&self, state: &GenericCommandState) -> Self {
         match self {
-            OperationAction::Script(script, handlers) => OperationAction::Script(
-                Self::inject_values_into_script(state, script),
-                handlers.clone(),
-            ),
-            OperationAction::BgScript(script, handlers) => OperationAction::BgScript(
-                Self::inject_values_into_script(state, script),
-                handlers.clone(),
-            ),
+            OperationAction::Script(script, handlers) => {
+                OperationAction::Script(script.inject_values(state), handlers.clone())
+            }
+            OperationAction::BgScript(script, handlers) => {
+                OperationAction::BgScript(script.inject_values(state), handlers.clone())
+            }
             OperationAction::Operation(operation_expr, optional_script, input, handlers) => {
                 let operation = state.inject_values_into_template(operation_expr);
                 let optional_script = optional_script
                     .as_ref()
-                    .map(|script| Self::inject_values_into_script(state, script));
+                    .map(|script| script.inject_values(state));
                 OperationAction::Operation(
                     operation,
                     optional_script,
@@ -366,13 +365,6 @@ impl OperationAction {
                 )
             }
             _ => self.clone(),
-        }
-    }
-
-    fn inject_values_into_script(state: &GenericCommandState, script: &ShellScript) -> ShellScript {
-        ShellScript {
-            command: state.inject_values_into_template(&script.command),
-            args: state.inject_values_into_parameters(&script.args),
         }
     }
