+++ crates/core/tedge_api/src/workflow/state.rs
@@ -2,6 +2,7 @@ use crate::mqtt_topics::Channel;
 use crate::mqtt_topics::EntityTopicId;
 use crate::mqtt_topics::MqttSchema;
 use crate::mqtt_topics::OperationType;
+use crate::substitution::Record;
 use crate::workflow::CommandId;
 use crate::workflow::ExitHandlers;
 use crate::workflow::OperationName;
@@ -284,65 +285,15 @@ impl GenericCommandState {
             o.insert(property.to_string(), value.into());
         }
     }
+}
 
-    /// Inject values extracted from the message payload into a script command line.
-    ///
-    /// - The script command is first tokenized using shell escaping rules.
-    ///   `/some/script.sh arg1 "arg 2" "arg 3"` -> ["/some/script.sh", "arg1", "arg 2", "arg 3"]
-    /// - Then each token matching `${x.y.z}` is substituted with the value pointed by the JSON path.
-    pub fn inject_values_into_parameters(&self, args: &[String]) -> Vec<String> {
-        args.iter()
-            .map(|arg| self.inject_values_into_template(arg))
-            .collect()
-    }
-
-    /// Inject values extracted from the message payload into a template string
-    ///
-    /// - Search the template string for path patterns `${...}`
-    /// - Replace all these paths by the value extracted from self using the paths
-    ///
-    /// `"prefix-${.payload.x}-separator-${.payload.y}-suffix"` is replaced by
-    /// `"prefix-X-separator-Y-suffix"` in a context where the payload is `{"x":"X", "y":"Y"}`
-    pub fn inject_values_into_template(&self, target: &str) -> String {
-        target
-            .split_inclusive('}')
-            .flat_map(|s| match s.find("${") {
-                None => vec![s],
-                Some(i) => {
-                    let (prefix, template) = s.split_at(i);
-                    vec![prefix, template]
-                }
-            })
-            .map(|s| self.replace_path_with_value(s))
-            .collect()
-    }
-
-    /// Replace a path pattern with the value extracted from the message payload using that path
-    ///
-    /// `${.payload}` -> the whole message payload
-    /// `${.payload.x}` -> the value of x if there is any in the payload
-    /// `${.payload.unknown}` -> `${.payload.unknown}` unchanged
-    /// `Not a path expression` -> `Not a path expression` unchanged
-    fn replace_path_with_value(&self, template: &str) -> String {
-        Self::extract_path(template)
-            .and_then(|path| self.extract_value(path))
-            .map(|v| json_as_string(&v))
-            .unwrap_or_else(|| template.to_string())
-    }
-
-    /// Extract a path  from a `${ ... }` expression
-    ///
-    /// Return None if the input is not a path expression
-    pub fn extract_path(input: &str) -> Option<&str> {
-        input.strip_prefix("${").and_then(|s| s.strip_suffix('}'))
-    }
-
+impl Record for GenericCommandState {
     /// Extract the JSON value pointed by a path from this command state
     ///
     /// Return None if the path contains unknown fields,
     /// with the exception that the empty string is returned for an unknown path below the `.payload`,
     /// the rational being that the payload object represents a free-form value.
-    pub fn extract_value(&self, path: &str) -> Option<Value> {
+    fn extract_value(&self, path: &str) -> Option<Value> {
         match path {
             "." => Some(json!({
                 "topic": self.topic.name,
@@ -364,7 +315,9 @@ impl GenericCommandState {
             }
         }
     }
+}
 
+impl GenericCommandState {
     /// Return the topic that uniquely identifies the command
     pub fn command_topic(&self) -> &String {
         &self.topic.name
@@ -591,16 +544,6 @@ fn json_excerpt<'a>(value: &'a Value, path: &'a str) -> Option<&'a Value> {
         None if path.is_empty() => Some(value),
         None => value.get(path),
         Some((key, path)) => value.get(key).and_then(|value| json_excerpt(value, path)),
-    }
-}
-
-fn json_as_string(value: &Value) -> String {
-    match value {
-        Value::Null => "null".to_string(),
-        Value::Bool(b) => b.to_string(),
-        Value::Number(n) => n.to_string(),
-        Value::String(s) => s.clone(),
-        _ => value.to_string(),
     }
 }
